requirements.txt
streamlit==1.39.0
langchain==0.3.20
langchain-community==0.3.19
langchain-groq==0.2.5
sqlalchemy==2.0.30
pandas==2.2.2
plotly==5.20.0
python-dotenv==1.0.1
